refactor to expose read_swc
