+++ src/napari_swc_editor/__init__.py
@@ -2,7 +2,7 @@ try:
     from ._version import version as __version__
 except ImportError:
     __version__ = "unknown"
-from ._reader import napari_get_reader
+from ._reader import napari_get_reader, read_swc
 from ._sample_data import make_sample_data
 from ._widget import (
     ExampleQWidget,
